refactor: reorganize requires updates functionality
Removed the `requires_test.go` file and integrated its tests into the new `requiresupdates` package. Introduced new files for managing compatibility checks and manifest updates, enhancing the structure and maintainability of the codebase. This change improves the organization of dependency management for integration packages, ensuring better clarity and separation of concerns.
