+++ internal/requiresupdates/compatibility_test.go
