+++ internal/requiresupdates/manifest_test.go
